backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==1.2.0
requests==2.31.0
python-dotenv==1.0.0
httpx==0.24.1
gunicorn==21.2.0
